[fix] change the place
