+++ server.js
@@ -25,7 +25,6 @@ const translation       = require('./utils/translation');
 const serverUtils       = require('./utils/server');
 const utilsModules      = require('./utils/modules');
 const utilsThemes       = require('./utils/themes');
-const NSErrors          = require('./utils/errors/NSErrors');
 const {
     middlewarePassport,
     expressErrorHandler,
@@ -127,50 +126,7 @@ const initServer = async () => {
         require('./services/cache').cacheSetting();
         const apiRouter = require('./routes').InitRoutes(express, server);
         if (dev) {
-            const reloadRouter  = express.Router();
-            let apiIndexInStack;
-            const numberOfStack = server._router.stack.length;
-            for (let i = 0; i < numberOfStack; i++) {
-                if (server._router.stack[i].regexp.toString().includes('api')
-        && server._router.stack[i].name === 'router'
-        && server._router.stack[(i + 1)].regexp.toString().includes('api')) {
-                    apiIndexInStack = i;
-                    break;
-                }
-            }
-            reloadRouter.route('/reloadAPI').get(async (req, res) => {
-                try {
-                    // Remove the matched middleware
-                    server._router.stack.splice(apiIndexInStack, 1);
-                    server._router.stack.splice(apiIndexInStack, 1);
-                    server._router.stack.splice(apiIndexInStack, 1);
-                    const newAPIRouter        = express.Router();
-                    const newAdminFrontRouter = express.Router();
-                    server.use('/api', newAPIRouter, reloadRouter, (req, res, next) => next(NSErrors.ApiNotFound));
-                    for (let i = 0; i < 3; i++) {
-                        const lastStack = server._router.stack[server._router.stack.length - 1];
-                        server._router.stack.pop();
-                        server._router.stack.splice(apiIndexInStack, 0, lastStack);
-                    }
-                    // we remove cache of /routes/* and /services/*
-                    for (const oneFolder of ['routes', 'services']) {
-                        const pathToFolder = path.join(global.appRoot, oneFolder);
-                        for (const oneLine in require.cache) {
-                            if (oneLine.startsWith(pathToFolder)) {
-                                delete require.cache[oneLine];
-                            }
-                        }
-                    }
-                    require('./routes').loadDynamicRoutes(newAPIRouter, newAdminFrontRouter);
-                    return res.json('ok');
-                } catch (errorInReload) {
-                    return res.json(errorInReload);
-                }
-            });
-            server.use('/api', reloadRouter);
-            const lastStack = server._router.stack[server._router.stack.length - 1];
-            server._router.stack.pop();
-            server._router.stack.splice(apiIndexInStack, 0, lastStack);
+            await serverUtils.hotReloadAPI(express, server);
         }
 
         await utilsModules.modulesLoadInitAfter(apiRouter, server, passport);
